drag and drop functionality tidied, data being retrieved from apis and rendered into drag/drop boxes
