+++ client/src/App.js
@@ -6,15 +6,13 @@ import ParentDashboard from './containers/parent/ParentDashboard';
 import StudentDashboard from './containers/student/StudentDashboard';
 import QuizPage from './components/student/QuizPage';
 import CompletionPage from './components/student/CompletionPage';
-import DnDPage from './components/student/DnDPage';
-import { DndProvider } from 'react-dnd'
-import { HTML5Backend } from 'react-dnd-html5-backend'
+import DnDPage from './components/student/DnD/DnDPage';
+
 
 
 function App() {
   return (
     <Router>    
-      <DndProvider backend={HTML5Backend}>
       <Navbar></Navbar>
       <main>
         <Switch>
@@ -26,7 +24,6 @@ function App() {
         <Route path='/student/:topic/completed' component={CompletionPage}/>
         </Switch>
       </main>
-      </DndProvider>
     </Router> 
     );
